requirements.txt
PyYAML==3.12
prometheus-client==0.4.2